# [RFC] mutate_locally: destroy mutations gently