--- a/db/batchlog_manager.cc
+++ b/db/batchlog_manager.cc
@@ -264,7 +264,7 @@ future<> db::batchlog_manager::replay_all_failed_batches(post_replay_cleanup cle
             mutation m(schema, key);
             auto now = service::client_state(service::client_state::internal_tag()).get_timestamp();
             m.partition().apply_delete(*schema, clustering_key_prefix::make_empty(), tombstone(now, gc_clock::now()));
-            return _qp.proxy().mutate_locally(m, tracing::trace_state_ptr(), db::commitlog::force_sync::no);
+            return _qp.proxy().mutate_locally(std::move(m), tracing::trace_state_ptr(), db::commitlog::force_sync::no);
         }).then([] { return make_ready_future<stop_iteration>(stop_iteration::no); });
     };