--- a/mutation/mutation.cc
+++ b/mutation/mutation.cc
@@ -44,6 +44,19 @@ mutation::data::data(schema_ptr&& schema, dht::decorated_key&& key, mutation_par
     , _p(std::move(mp))
 { }
 
+future<> mutation::data::clear_gently() noexcept {
+    _schema = nullptr;
+    return repeat([this] { return _p.clear_gently(nullptr); });
+}
+
+future<> mutation::clear_gently() noexcept {
+    if (auto p = std::exchange(_ptr, nullptr)) {
+        auto f = p->clear_gently();
+        return f.available() ? std::move(f) : f.finally([p = std::move(p)] {});
+    }
+    return make_ready_future();
+}
+
 void mutation::set_static_cell(const column_definition& def, atomic_cell_or_collection&& value) {
     partition().static_row().apply(def, std::move(value));
 }