--- a/service/storage_proxy.cc
+++ b/service/storage_proxy.cc
@@ -3074,6 +3074,13 @@ future<> apply_on_shards(const locator::effective_replication_map_ptr& erm, cons
     return seastar::parallel_for_each(shards, std::move(apply));
 }
 
+void storage_proxy::destroy_mutation_gently(mutation&& m) {
+    // the gate is closed on shutdown
+    (void)with_gate(_async_gate, [m = std::move(m)] () mutable {
+        return m.clear_gently();
+    });
+}
+
 future<>
 storage_proxy::mutate_locally(const mutation& m, tracing::trace_state_ptr tr_state, db::commitlog::force_sync sync, clock_type::time_point timeout, smp_service_group smp_grp, db::per_partition_rate_limit::info rate_limit_info) {
     auto erm = _db.local().find_column_family(m.schema()).get_effective_replication_map();
@@ -6832,6 +6839,9 @@ future<> storage_proxy::drain_on_shutdown() {
     return async([this] {
         cancel_write_handlers([] (const abstract_write_response_handler&) { return true; });
         _hints_resource_manager.stop().get();
+        if (!_async_gate.is_closed()) {
+            _async_gate.close().get();
+        }
     });
 }