--- a/service/storage_proxy.cc
+++ b/service/storage_proxy.cc
@@ -3434,14 +3434,14 @@ storage_proxy::hint_to_dead_endpoints(response_id_type id, db::consistency_level
     }
 }
 
-template<typename Range, typename CreateWriteHandler>
-future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mutate_prepare(Range&& mutations, db::consistency_level cl, db::write_type type, service_permit permit, CreateWriteHandler create_handler) {
+template<std::ranges::range Range>
+future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mutate_prepare_and_consume(Range&& mutations, db::consistency_level cl, db::write_type type, service_permit permit, std::function<result<storage_proxy::response_id_type>(std::ranges::range_value_t<Range>&&, db::consistency_level, db::write_type, service_permit)> create_handler) {
     // apply is used to convert exceptions to exceptional future
-    return futurize_invoke([this] (Range&& mutations, db::consistency_level cl, db::write_type type, service_permit permit, CreateWriteHandler create_handler) {
+    return futurize_invoke([this] (Range&& mutations, db::consistency_level cl, db::write_type type, service_permit permit, auto create_handler) {
         unique_response_handler_vector ids;
         ids.reserve(std::distance(std::begin(mutations), std::end(mutations)));
         for (auto&& m : mutations) {
-            auto r_handler = create_handler(m, cl, type, permit);
+            auto r_handler = create_handler(std::move(m), cl, type, permit);
             if (!r_handler) {
                 return make_ready_future<result<unique_response_handler_vector>>(std::move(r_handler).as_failure());
             }
@@ -3451,9 +3451,33 @@ future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mut
     }, std::forward<Range>(mutations), cl, type, std::move(permit), std::move(create_handler));
 }
 
-template<typename Range>
-future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mutate_prepare(Range&& mutations, db::consistency_level cl, db::write_type type, tracing::trace_state_ptr tr_state, service_permit permit, db::allow_per_partition_rate_limit allow_limit) {
-    return mutate_prepare<>(std::forward<Range>(mutations), cl, type, std::move(permit), [this, tr_state = std::move(tr_state), allow_limit] (const std::ranges::range_value_t<Range>& m, db::consistency_level cl, db::write_type type, service_permit permit) mutable {
+template<std::ranges::range Range>
+future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mutate_prepare_and_consume(Range&& mutations, db::consistency_level cl, db::write_type type, tracing::trace_state_ptr tr_state, service_permit permit, db::allow_per_partition_rate_limit allow_limit) {
+    return mutate_prepare_and_consume(std::forward<Range>(mutations), cl, type, std::move(permit), [this, tr_state = std::move(tr_state), allow_limit] (std::ranges::range_value_t<Range>&& m, db::consistency_level cl, db::write_type type, service_permit permit) mutable {
+        return create_write_response_handler(std::move(m), cl, type, tr_state, std::move(permit), allow_limit);
+    });
+}
+
+template<std::ranges::range Range>
+future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mutate_prepare(const Range& mutations, db::consistency_level cl, db::write_type type, service_permit permit, std::function<result<storage_proxy::response_id_type>(const std::ranges::range_value_t<Range>&, db::consistency_level, db::write_type, service_permit)> create_handler) {
+    // apply is used to convert exceptions to exceptional future
+    return futurize_invoke([this] (const Range& mutations, db::consistency_level cl, db::write_type type, service_permit permit, auto create_handler) {
+        unique_response_handler_vector ids;
+        ids.reserve(std::distance(std::begin(mutations), std::end(mutations)));
+        for (const auto& m : mutations) {
+            auto r_handler = create_handler(m, cl, type, permit);
+            if (!r_handler) {
+                return make_ready_future<result<unique_response_handler_vector>>(std::move(r_handler).as_failure());
+            }
+            ids.emplace_back(*this, std::move(r_handler).value());
+        }
+        return make_ready_future<result<unique_response_handler_vector>>(std::move(ids));
+    }, mutations, cl, type, std::move(permit), std::move(create_handler));
+}
+
+template<std::ranges::range Range>
+future<result<storage_proxy::unique_response_handler_vector>> storage_proxy::mutate_prepare(const Range& mutations, db::consistency_level cl, db::write_type type, tracing::trace_state_ptr tr_state, service_permit permit, db::allow_per_partition_rate_limit allow_limit) {
+    return mutate_prepare<>(mutations, cl, type, std::move(permit), [this, tr_state = std::move(tr_state), allow_limit] (const std::ranges::range_value_t<Range>& m, db::consistency_level cl, db::write_type type, service_permit permit) mutable {
         return create_write_response_handler(m, cl, type, tr_state, std::move(permit), allow_limit);
     });
 }
@@ -3784,7 +3808,7 @@ storage_proxy::mutate_internal(Range mutations, db::consistency_level cl, bool c
     utils::latency_counter lc;
     lc.start();
 
-    return mutate_prepare(mutations, cl, type, tr_state, std::move(permit), allow_limit).then(utils::result_wrap([this, cl, timeout_opt, tracker = std::move(cdc_tracker),
+    return mutate_prepare_and_consume(std::move(mutations), cl, type, tr_state, std::move(permit), allow_limit).then(utils::result_wrap([this, cl, timeout_opt, tracker = std::move(cdc_tracker),
             tr_state] (storage_proxy::unique_response_handler_vector ids) mutable {
         register_cdc_operation_result_tracker(ids, tracker);
         return mutate_begin(std::move(ids), cl, tr_state, timeout_opt);
@@ -3946,7 +3970,7 @@ storage_proxy::mutate_atomically_result(std::vector<mutation> mutations, db::con
         }
 
         future<result<>> send_batchlog_mutation(mutation m, db::consistency_level cl = db::consistency_level::ONE) {
-            return _p.mutate_prepare<>(std::array<mutation, 1>{std::move(m)}, cl, db::write_type::BATCH_LOG, _permit, [this] (const mutation& m, db::consistency_level cl, db::write_type type, service_permit permit) {
+            return _p.mutate_prepare(std::array<mutation, 1>{std::move(m)}, cl, db::write_type::BATCH_LOG, _permit, [this] (const mutation& m, db::consistency_level cl, db::write_type type, service_permit permit) {
                 return _p.create_write_response_handler(_ermp, cl, type, std::make_unique<shared_mutation>(m), _batchlog_endpoints, {}, {}, _trace_state, _stats, std::move(permit), std::monostate(), is_cancellable::no);
             }).then(utils::result_wrap([this, cl] (unique_response_handler_vector ids) {
                 _p.register_cdc_operation_result_tracker(ids, _cdc_tracker);
@@ -4071,9 +4095,9 @@ future<> storage_proxy::send_to_endpoint(
         // and to apply backpressure.
         timeout = clock_type::now() + 5min;
     }
-    return mutate_prepare(std::array{std::move(m)}, cl, type, /* does view building should hold a real permit */ empty_service_permit(),
+    return mutate_prepare_and_consume(std::array{std::move(m)}, cl, type, /* does view building should hold a real permit */ empty_service_permit(),
             [this, tr_state, erm = std::move(ermp), target = std::array{target}, pending_endpoints = std::move(pending_endpoints), &stats, cancellable] (
-                std::unique_ptr<mutation_holder>& m,
+                std::unique_ptr<mutation_holder>&& m,
                 db::consistency_level cl,
                 db::write_type type, service_permit permit) mutable {
         inet_address_vector_replica_set targets;