change the function signatures of mutate_locally
and create_write_response_handler functions to accept
the mutation(s) as rvalue refrences so they can be
destroyed gently in the background and being frozen
via the following helpers:
- freeze_mutation_and_schema_and_destroy_mutation
- freeze_and_destroy_mutation
- destroy_mutation_gently

Destroying the mutation in the background is done
to prevent reactor stalls when either large mutations
or a vector of (large) mutations are destroyed in the
mutate_locally path, on the return path.

Fixes #12999