@bhalevy 2023-02-27
It appears like with destroying write mutations gently in the background, memory pressure is built more intensively and we see significant stalls in memtable::apply.
https://github.com/scylladb/scylladb/pull/13004/commits/3d55725e43adffd8fdbcc7d1eae3d80291dad7e0 deals with these stalls by throwing bad_alloc which works in the test, but I'm not sure if it's the best way to go :-/