@scylladb-promoter 2023-02-27
CI state **`FAILURE`** - https://jenkins.scylladb.com/job/releng/job/Scylla-CI/4619/