@bhalevy 2023-02-27
> Does this fix a problem? mutations in mutate_locally are typically small, as they are the result of INSERT or UPDATE statements.
>
> I'm worried about performance reduction due to the change.

There's a 30-35ms stall reported in #12999 fixed by the first patches in this series.

The fix defers the destruction of mutations to the background so I assume it would actually improve burst performance, but with a steady high write workload, further testing revealed the deferring the destruction of mutations to the background increases the memory pressure on memtable apply and log allocation, causing painful stalls there.
Those stalls are addressed in the last patch in this series, but it will turn those stalls to spurious bad_alloc exceptions under high write workload.  I'm not sure if that is desired.  It's a very poor way to apply back pressure towards the client.