@bhalevy 2023-02-27
For completeness, here's the full list of stalls without  [3d55725](https://github.com/scylladb/scylladb/pull/13004/commits/3d55725e43adffd8fdbcc7d1eae3d80291dad7e0)

[stalls-report.txt](https://github.com/scylladb/scylladb/files/10840163/stalls-report.txt)
[stalls.txt](https://github.com/scylladb/scylladb/files/10840164/stalls.txt)