@tgrabiec 2023-02-27
This is in the schema merge path, right? 

It doesn't go through storage_proxy.