@bhalevy 2023-02-27
> This is in the schema merge path, right?
> 
> It doesn't go through storage_proxy.

#12999 happens in schema_tables store_column_mapping in the return path, after calling `proxy.local().mutate_locally`.
Then the vector of mutations is destroyed.