@avikivity 2023-02-27
Please check for performance regressions with perf_simple_query and add to the cover letter.