@tgrabiec 2023-02-27
Actually, we need them to be applied atomically, so we cannot do that incrementally. The current code seems buggy to me. If the node crashes while storing, we may end up with an incomplete column mapping.