@tgrabiec 2023-02-27
Btw, `memtable::apply()` is not prepared to apply with deferring in the middle. Fixing this only at storage_proxy is not enough to avoid stalls with huge mutations.