@bhalevy 2024-11-17
retracting this PR as https://github.com/scylladb/scylladb/issues/12999 was addressed
and this change causes slight write perf degradation.
I'll just push the latest rebase and performance results for reference