@bhalevy 2024-11-17
Before:
```
==> perf-simple-query.before <==
random-seed=2722541068
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=read, frontend=cql, query_single_key=no, counters=no}
Disabling auto compaction
Creating 10000 partitions...
206376.41 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41690 insns/op,   21972 cycles/op,        0 errors)
214662.52 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41691 insns/op,   21409 cycles/op,        0 errors)
212039.11 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41710 insns/op,   21648 cycles/op,        0 errors)
214440.56 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41673 insns/op,   21551 cycles/op,        0 errors)
211271.77 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41703 insns/op,   21843 cycles/op,        0 errors)

         throughput: mean=211758.07 standard-deviation=3350.72 median=212039.11 median-absolute-deviation=2682.49 maximum=214662.52 minimum=206376.41
instructions_per_op: mean=41693.48 standard-deviation=14.34 median=41690.94 median-absolute-deviation=9.91 maximum=41710.34 minimum=41673.00
  cpu_cycles_per_op: mean=21684.58 standard-deviation=225.37 median=21648.46 median-absolute-deviation=158.75 maximum=21971.88 minimum=21408.58

==> perf-simple-query-write.before <==
random-seed=3344697609
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=write, frontend=cql, query_single_key=no, counters=no}
Disabling auto compaction
168265.75 tps ( 59.4 allocs/op,  16.2 logallocs/op,  14.3 tasks/op,   51890 insns/op,   25502 cycles/op,        0 errors)
173458.32 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   51905 insns/op,   25662 cycles/op,        0 errors)
174286.95 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   51936 insns/op,   25632 cycles/op,        0 errors)
174243.49 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   51916 insns/op,   25597 cycles/op,        0 errors)
173125.44 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   51920 insns/op,   25718 cycles/op,        0 errors)

         throughput: mean=172675.99 standard-deviation=2515.78 median=173458.32 median-absolute-deviation=1567.51 maximum=174286.95 minimum=168265.75
instructions_per_op: mean=51913.36 standard-deviation=16.89 median=51915.98 median-absolute-deviation=8.20 maximum=51935.73 minimum=51890.37
  cpu_cycles_per_op: mean=25622.03 standard-deviation=80.35 median=25631.64 median-absolute-deviation=40.16 maximum=25717.51 minimum=25502.21
```

After:
```
==> perf-simple-query.after <==
random-seed=3594221815
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=read, frontend=cql, query_single_key=no, counters=no}
Disabling auto compaction
Creating 10000 partitions...
179673.20 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41661 insns/op,   23051 cycles/op,        0 errors)
200163.48 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41693 insns/op,   22191 cycles/op,        0 errors)
211299.51 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41673 insns/op,   21891 cycles/op,        0 errors)
206437.21 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41725 insns/op,   22404 cycles/op,        0 errors)
209394.55 tps ( 63.1 allocs/op,   0.0 logallocs/op,  14.1 tasks/op,   41708 insns/op,   22084 cycles/op,        0 errors)

         throughput: mean=201393.59 standard-deviation=12853.69 median=206437.21 median-absolute-deviation=8000.96 maximum=211299.51 minimum=179673.20
instructions_per_op: mean=41691.73 standard-deviation=25.72 median=41692.64 median-absolute-deviation=19.01 maximum=41724.57 minimum=41660.95
  cpu_cycles_per_op: mean=22324.19 standard-deviation=446.63 median=22191.30 median-absolute-deviation=240.46 maximum=23050.96 minimum=21890.79

==> perf-simple-query-write.after <==
random-seed=2731171770
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=write, frontend=cql, query_single_key=no, counters=no}
Disabling auto compaction
164265.17 tps ( 60.4 allocs/op,  16.2 logallocs/op,  14.3 tasks/op,   52341 insns/op,   26099 cycles/op,        0 errors)
161311.11 tps ( 60.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   52417 insns/op,   27516 cycles/op,        0 errors)
169056.95 tps ( 60.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   52436 insns/op,   26243 cycles/op,        0 errors)
171090.24 tps ( 60.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   52434 insns/op,   25989 cycles/op,        0 errors)
171401.70 tps ( 60.3 allocs/op,  16.0 logallocs/op,  14.2 tasks/op,   52442 insns/op,   25921 cycles/op,        0 errors)

         throughput: mean=167425.03 standard-deviation=4451.84 median=169056.95 median-absolute-deviation=3665.21 maximum=171401.70 minimum=161311.11
instructions_per_op: mean=52414.08 standard-deviation=41.81 median=52433.79 median-absolute-deviation=22.31 maximum=52442.39 minimum=52341.29
  cpu_cycles_per_op: mean=26353.50 standard-deviation=660.83 median=26098.58 median-absolute-deviation=364.12 maximum=27515.50 minimum=25921.44
```